test(ops): harden Railway drift regression coverage (#6091)
